requirements.txt
pyyaml==6.0.2
openai==1.51.2
requests==2.31.0
prompt_toolkit==3.0.38
rich==13.4.2
pydantic==2.6.4